Align the topics page to the designs (#207)
* Align the topics page to the designs

Update the Topics page header to be in sync with the designs. The warning icon will always show zero as for the moment we don't have the count of topics that are in-sync and under replicated.

Connect the search input to search in the topic names.

Add a switch to show/hide internal topics (hide is the default).
not in sync. Part of #165.

Add a status column to show the in-sync or under replicated state of a topic. This is done by looking for a replica amongs all partitions that hold the topic that is flagged as not in sync.

Reorder the columns.

Change the "Edit properties" action to "Edit configuration". Part of #142.

Change the column name "Messages" to "Record count".

Add the right empty states for when there are no topics and for when a search doesn't produce results.

* Add an option to search by topic id

* Update icons

* Minor tweaks